move MdSpanArray in an extra header file
